requirements.txt
boto3==1.33.2
botocore==1.33.2
s3transfer==0.8.1
psycopg2-binary
